fixing nested message types
note: i found a bug in scalac's java parser that causes attempts to compile java protoc output with nested message types to fail.  it is caused by having a class with an inner class called Builder and another inner class that also has an inner class called Builder.  if you want to use nested message types with scala you have to compile your java output separately first and put the jar or .class in the classpath.